requirements.txt
python-dotenv>=1.0.0
langchain-google-genai>=0.0.5
langchain-chroma>=0.1.0
langchain-community>=0.0.20
unstructured>=0.10.0
chromadb>=0.4.0

# Optional but recommended for PDF processing
pdf2image>=1.16.0
pdfminer.six>=20221105
